# Client-Backend/src/swagger.yaml
openapi: 3.0.0
info:
  title: Home Tuition Backend API
  description: |
    This is the backend API for the **Home Tuition Platform**, designed to manage tutors, students, and class schedules efficiently. 
    
    The API provides endpoints for user authentication, tutor registration, student enrollment, and class management. 
    
    We encourage contributions to improve and extend the API. Feel free to explore, suggest changes, or enhance the features.
    
    Some useful links:
    - [Backend Repository](https://github.com/razrahul/Client-Backend)
    - [Frontend UI](https://contact-fromdata.netlify.app)
    
  termsOfService: http://swagger.io/terms/
  contact:
    name: Rahul Kumar
    email: razrahuleee@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://client-backend-ev0b.onrender.com/api/v1
    description: Production Server
  - url: http://localhost:5500/api/v1
    description: Local Development Server
components:
  securitySchemes:
    noauthAuth:
      type: http
      scheme: noauth
...
paths:
...
  /subject:
    get:
      tags:
        - Subject
      summary: get All Subject
      security:
        - noauthAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - Subject
      summary: create subject
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: computer
      security:
        - noauthAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...